added Metrics Gui
libary source :
https://github.com/GameTechDev/MetricsGui

keep in mind when using this gui, in the start game screen before you load into your char the frame rate will be very high and the frame time will be very low. this is a expected outcome as the game doesnt have very much code it is running at this point.